+++ src/templates/vite-react/src/routes/Layout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import { ErrorBoundary } from 'react-error-boundary';
 import PageError from '@/components/app/PageError';
-import { Toaster } from '@/components/ui/toaster';
+import { Toaster } from '@/components/ui/Toaster';
 
 const Layout = () => {
   return (
